QC UI changes
