+++ layouts/AppTopbar.vue
@@ -88,11 +88,6 @@ const menuBar = computed(() => {
             url: "/hermes/",
         },
         {
-            label: "QC Reports",
-            icon: "bi-file-earmark-check",
-            url: "/hermes/qc/",
-        },
-        {
             label: "Meta-Analyses",
             icon: "bi-file-earmark-bar-graph",
             url: "/hermes/ma/",
